[Refactor] 프로토콜 이름 수정 FirebaseAuthService -> AuthService
인증서버가 교체할 수 있게 설계 하는게 클린아키텍처를 구현하는 목적에 더 맞다는 팀원들의 합의가 있어서 바꿨다.
