+++ NearTalk/NearTalk/Infrastructure/Network/FirebaseKey.swift
@@ -5,7 +5,6 @@
 //  Created by 고병학 on 2022/11/17.
 //
 
-import FirebaseFirestore
 import Foundation
 
 enum FirebaseKey {
