+++ NearTalk/NearTalk/Infrastructure/Network/StorageService.swift
@@ -10,6 +10,7 @@ import Foundation
 import RxSwift
 
 protocol StorageService {
+    /// Firebase storage에 저장하고 저장된 path(URL x)를 반환한다.
     func uploadData(data: Data, fileName: String, dataType: FirebaseKey.Storage) -> Single<String>
     func downloadURL(for path: String) -> Single<URL>
 }
