+++ NearTalk/NearTalk/Presentation/LaucnScreen/ViewModel/LaunchScreenViewModel.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import RxCocoa
 import RxSwift
 
 struct LaunchScreenViewModelActions {
